Artificial Neural Network

This Class represents an Artificial Neural Network model object and hosts all the information related to the model.

Usage

# S4 method for class 'ANN'
show(object)

Arguments

object

ANN object

Details

See nnet for the meaning of the slots.

Slots

size

integer. Number of the units in the hidden layer.

decay

numeric. Weight decay.

rang

numeric. Initial random weights.

maxit

integer. Maximum number of iterations.

model

nnet. The nnet model object.

Author

Sergio Vignali