AICc

Compute the Akaike Information Criterion corrected for small samples size (Warren and Seifert, 2011).

aicc(model, env, parallel = FALSE)

Arguments

model
SDMmodel object.

env
stack containing the environmental variables.

parallel
deprecated.

Value

The computed AICc

Details

The function is available only for Maxent and Maxnet methods.

References

Warren D.L., Seifert S.N., (2011). Ecological niche modeling in Maxent: the importance of model complexity and the performance of model selection criteria. Ecological Applications, 21(2), 335–342.

See also

Author

Sergio Vignali

Examples

# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")

# Train a model
model <- train(method = "Maxnet", data = data, fc = "l")

# Compute the AICc
aicc(model, predictors)