Merge SWD Objects

Merge two SWD objects.

Usage

mergeSWD(swd1, swd2, only_presence = FALSE)

Arguments

swd1

SWD object.

swd2

SWD object.

only_presence

logical If TRUE only for the presence locations are merged and the absence/background locations are taken only from the swd1 object.

Value

The merged SWD object.

Details

In case the two SWD objects have different columns, only the common columns are used in the merged object.

The SWD object is created in a way that the presence locations are always before than the absence/background locations.

Author

Sergio Vignali

Examples

# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd",
                    full.names = TRUE)

predictors <- terra::rast(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species",
                   p = p_coords,
                   a = bg_coords,
                   env = predictors,
                   categorical = "biome")
#> ℹ Extracting predictor information for presence locations
#> ✔ Extracting predictor information for presence locations [20ms]
#> 
#> ℹ Extracting predictor information for absence/background locations
#> ✔ Extracting predictor information for absence/background locations [46ms]
#> 

# Split only presence locations in training (80%) and testing (20%) datasets
datasets <- trainValTest(data,
                         test = 0.2,
                         only_presence = TRUE)
train <- datasets[[1]]
test <- datasets[[2]]

# Merge the training and the testing datasets together
merged <- mergeSWD(train,
                   test,
                   only_presence = TRUE)

# Split presence and absence locations in training (80%) and testing (20%)
datasets
#> [[1]]
#> 
#> ── Object of class: <SWD> ──
#> 
#> ── Info 
#> • Species: Virtual species
#> • Presence locations: 320
#> • Absence locations: 5000
#> 
#> ── Variables 
#> • Continuous: "bio1", "bio12", "bio16", "bio17", "bio5", "bio6", "bio7", and
#> "bio8"
#> • Categorical: "biome"
#> 
#> [[2]]
#> 
#> ── Object of class: <SWD> ──
#> 
#> ── Info 
#> • Species: Virtual species
#> • Presence locations: 80
#> • Absence locations: 5000
#> 
#> ── Variables 
#> • Continuous: "bio1", "bio12", "bio16", "bio17", "bio5", "bio6", "bio7", and
#> "bio8"
#> • Categorical: "biome"
#> 
datasets <- trainValTest(data,
                         test = 0.2)
train <- datasets[[1]]
test <- datasets[[2]]

# Merge the training and the testing datasets together
merged <- mergeSWD(train, test)