Model Report

Make a report that shows the main results.

Usage

modelReport(
  model,
  folder,
  test = NULL,
  type = NULL,
  response_curves = FALSE,
  only_presence = FALSE,
  jk = FALSE,
  env = NULL,
  clamp = TRUE,
  permut = 10,
  verbose = TRUE
)

Arguments

model

SDMmodel object.

folder

character. The name of the folder in which to save the output. The folder is created in the working directory.

test

SWD object with the test locations.

type

character. The output type used for "Maxent" and "Maxnet" methods, possible values are "cloglog" and "logistic".

response_curves

logical, if TRUE it plots the response curves in the html output.

only_presence

logical, if TRUE it uses only the range of the presence location for the marginal response.

jk

logical, if TRUE it runs the jackknife test.

env

rast. If provided it computes and adds a prediction map to the output.

clamp

logical for clumping during prediction, used for response curves and for the prediction map.

permut

integer. Number of permutations.

verbose

logical, if TRUE prints informative messages.

Details

The function produces a report similar to the one created by MaxEnt software. See terra documentation to see how to pass factors.

Author

Sergio Vignali

Examples

# If you run the following examples with the function example(),
# you may want to set the argument ask like following: example("modelReport",
# ask = FALSE)
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd",
                    full.names = TRUE)

predictors <- terra::rast(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species",
                   p = p_coords,
                   a = bg_coords,
                   env = predictors,
                   categorical = "biome")
#> ℹ Extracting predictor information for presence locations
#> ✔ Extracting predictor information for presence locations [22ms]
#> 
#> ℹ Extracting predictor information for absence/background locations
#> ✔ Extracting predictor information for absence/background locations [49ms]
#> 

# Split presence locations in training (80%) and testing (20%) datasets
datasets <- trainValTest(data,
                         test = 0.2,
                         only_presence = TRUE)
train <- datasets[[1]]
test <- datasets[[2]]

# Train a model
model <- train(method = "Maxnet",
               data = train,
               fc = "lq")

# Create the report
if (FALSE) { # \dontrun{
modelReport(model,
            type = "cloglog",
            folder = "my_folder",
            test = test,
            response_curves = TRUE,
            only_presence = TRUE,
            jk = TRUE,
            env = predictors,
            permut = 2)} # }