Plot Response Curve

Plot the Response Curve of the given environmental variable.

plotResponse(model, var, type, marginal = FALSE, fun = mean,
  clamp = TRUE, rug = FALSE, color = "red")

Arguments

model
SDMmodel or SDMmodelCV object.

var
character. Name of the variable to be plotted.

type
character. Output type, see predict,Maxent-method for Maxent models or predict.maxnet for Maxnet models.

marginal
logical, if TRUE it plots the marginal response curve, default is FALSE.

fun
function used to compute the level of the other variables for marginal curves, possible values are mean and median, default is mean.

clamp
logical for clumping during prediction, default is TRUE.

rug
logical, if TRUE it adds the rug plot for the presence and absence/background locations, available only for continuous variables, default is FALSE.

color
The color of the curve, default is "red".

Value

The ggplot2 object

Details

Note that fun is not a character argument, you must use mean and not "mean".

Examples

# NOT RUN {
plotResponse(model, var = "bio12", marginal = T, fun = median, rug = T)
# }