Predict Maxent

Predict the output for a new dataset from a trained Maxent model.

Usage

# S4 method for class 'Maxent'
predict(object, data, type = c("cloglog", "logistic", "raw"), clamp = TRUE)

Arguments

object

Maxent object.

data

data.frame with the data for the prediction.

type

character. MaxEnt output type, possible values are "cloglog", "logistic" and "raw".

clamp

logical for clumping during prediction.

Value

A vector with the prediction

Details

Used by the predict,SDMmodel-method, not exported.

The function performs the prediction in R without calling the MaxEnt Java software. This results in a faster computation for large datasets and might result in a slightly different output compared to the Java software.

References

Wilson P.D., (2009). Guidelines for computing MaxEnt model output values from a lambdas file.

Author

Sergio Vignali