Predict

Predict the output for a new dataset from a trained SDMmodel model.

# S4 method for SDMmodel
predict(object, data, type, clamp = TRUE,
  filename = "", format = "GTiff", extent = NULL, parallel = FALSE,
  progress = "", ...)

Arguments

object
SDMmodel object.

data
data.frame, SWD, stack.

type
character. Output type, see predict,Maxent-method for Maxent models or predict.maxnet for Maxnet models.

clamp
logical for clumping during prediction, default is TRUE.

filename
character. Output file name for the prediction map, if provided the output is saved in a file.

format
character. The output format, see writeRaster for all the options, default is "GTiff".

extent
Extent object, if provided it restricts the prediction to the given extent, default is NULL.

parallel
logical to use parallel computation during prediction, default is FALSE.

progress
character to display a progress bar: "text", "window" or "" (default) for no progress bar.

...
Additional arguments to pass to the writeRaster function.

Value

A vector with the prediction or a Raster object if data is a raster stack.

Details

Parallel computation increases the speed only for big datasets due to the time necessary to create the cluster. For Maxent models the function performs the prediction in R without calling the MaxEnt java software. This results is a faster computation for large datasets.

References

Wilson P.D., (2009). Guidelines for computing MaxEnt model output values from a lambdas file.

Examples

# NOT RUN {
predict(model, predictors, parallel = TRUE)
# }