Tune Feature Combination

The function iterates among different Feature Classes combinations and for each of them performs trains a different MaxEnt model.

tuneFC(model, fcs, metric = c("auc", "tss", "aicc"), test = NULL,
  env = NULL, parallel = FALSE)

Arguments

model
SDMmodel or SWDmodelCV object.

fcs
vector with all the Feature Classes combination to be tested.

metric
character. The metric used to evaluate the models, possible values are: "auc", "tss" and "aicc", default is "auc".

test
SWD. Test dataset used to evaluate the model, not used with aicc and SDMmodelCV objects, default is NULL.

env
stack containing the environmental variables, used only with "aicc", default is NULL.

parallel
logical, if TRUE it uses parallel computation, deafult is FALSE. Used only with AICc.

Value

A SDMtune object.

Details

You need package snow to use parallel computation and rgdal to save the prediction in a raster file. Parallel computation increases the speed only for big datasets due to the time necessary to create the cluster.

See also

Examples

# NOT RUN {
fc_test <- tuneFC(model, fcs = c("l", "lq", "lqp"), metric = "aicc",
env = predictors, parallel = T)
# }