Variable Selection

The function performs a data-driven variable selection. Starting from the provided model it iterates through all the variables starting from the one with the highest contribution (permutation importance). If the variable is correlated with other variables (according to the given method and threshold) it performs a Jackknife test and among the correlated variables it removes the one that results in the best performing model when removed (according to the given metric using the test dataset). The process is repeated untill the remaining variables are not highly correlated anymore.

varSel(model, bg4cor, metric = c("auc", "tss", "aicc"), test = NULL,
  env = NULL, parallel = FALSE, reg = 0.1, method = "spearman",
  cor_th = 0.7, permut = 10)

Arguments

model
SDMmodel or SDMmodelCV object.

bg4cor
SWD object. Background locations used to test the correlation between environmental variables.

metric
character. The metric used to evaluate the models, possible values are: "auc", "tss" and "aicc", default is "auc".

test
SWD. Test dataset used to evaluate the model, not used with aicc and SDMmodelCV objects, default is NULL.

env
stack containing the environmental variables, used only with "aicc", default is NULL.

parallel
logical, if TRUE it uses parallel computation, deafult is FALSE. Used only with AICc.

reg
integer. The value of the regularization paramiter to use during computation, default is 0.1, see details.

method
character. The method used to comput the correlation matrix, default "spearman".

cor_th
numeric. The correlation threshold used to select highly correlated variables, default is 0.7.

permut
integer. Number of permutations, default is 10.

Value

The model with trained using the selected variables.

Details

You need package snow to use parallel computation. Parallel computation increases the speed only for big datasets due to the time necessary to create the cluster. For Maxnet using a reg lower than 0.1 the model won't converge! We should write something more... I will refer to our paper for the explanations...

Examples

# NOT RUN {
varSel(model, bg, metric = "auc")
# }