Grid Search

Given a set of possible hyperparameter values, the function trains models with all the possible combinations of hyperparameters.

gridSearch(
  model,
  hypers,
  metric,
  test = NULL,
  env = NULL,
  save_models = TRUE,
  interactive = TRUE
)

Arguments

model

SDMmodel or SDMmodelCV object.

hypers

named list containing the values of the hyperparameters that should be tuned, see details.

metric

character. The metric used to evaluate the models, possible values are: "auc", "tss" and "aicc".

test

SWD object. Testing dataset used to evaluate the model, not used with aicc and SDMmodelCV objects, default is NULL.

env

stack containing the environmental variables, used only with "aicc", default is NULL.

save_models

logical, if FALSE the models are not saved and the output contains only a data frame with the metric values for each hyperparameter combination. Default is TRUE, set it to FALSE when there are many combinations to avoid R crashing for memory overload.

interactive

logical, if FALSE the interactive chart is not created.

Value

SDMtune object.

Details

To know which hyperparameters can be tuned you can use the output of the function getTunableArgs. Hyperparameters not included in the hypers argument take the value that they have in the passed model.

An interactive chart showing in real-time the steps performed by the algorithm is displayed in the Viewer pane.

See also

Author

Sergio Vignali

Examples

# \donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")
#> ℹ Extracting predictor information for presence locations
#> ✔ Extracting predictor information for presence locations [13ms]
#> 
#> ℹ Extracting predictor information for absence/background locations
#> ✔ Extracting predictor information for absence/background locations [12ms]
#> 

# Split presence locations in training (80%) and testing (20%) datasets
datasets <- trainValTest(data, test = 0.2, only_presence = TRUE)
train <- datasets[[1]]
test <- datasets[[2]]

# Train a model
model <- train(method = "Maxnet", data = train, fc = "l")

# Define the hyperparameters to test
h <- list(reg = 1:2, fc = c("lqp", "lqph"))

# Run the function using the AUC as metric
output <- gridSearch(model, hypers = h, metric = "auc", test = test)
#> Grid Search  ■■■■■■■■■■■■■                     40% | ETA:  4s - 00:00:2.7
#> Grid Search  ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s - 00:00:4.9
#> 
output@results
#>     fc reg train_AUC  test_AUC   diff_AUC
#> 1  lqp   1 0.8634203 0.8519775 0.01144281
#> 2 lqph   1 0.8727191 0.8567250 0.01599406
#> 3  lqp   2 0.8597322 0.8486725 0.01105969
#> 4 lqph   2 0.8657109 0.8501575 0.01555344
output@models
#> [[1]]
#> Object of class SDMmodel 
#> Method: Maxnet 
#> 
#> Species: Virtual species 
#> Presence locations: 320 
#> Absence locations: 5000 
#> 
#> Model configurations:
#> --------------------
#> fc: lqp
#> reg: 1
#> 
#> Variables:
#> ---------
#> Continuous: bio1 bio12 bio16 bio17 bio5 bio6 bio7 bio8 
#> Categorical: biome
#> [[2]]
#> Object of class SDMmodel 
#> Method: Maxnet 
#> 
#> Species: Virtual species 
#> Presence locations: 320 
#> Absence locations: 5000 
#> 
#> Model configurations:
#> --------------------
#> fc: lqph
#> reg: 1
#> 
#> Variables:
#> ---------
#> Continuous: bio1 bio12 bio16 bio17 bio5 bio6 bio7 bio8 
#> Categorical: biome
#> [[3]]
#> Object of class SDMmodel 
#> Method: Maxnet 
#> 
#> Species: Virtual species 
#> Presence locations: 320 
#> Absence locations: 5000 
#> 
#> Model configurations:
#> --------------------
#> fc: lqp
#> reg: 2
#> 
#> Variables:
#> ---------
#> Continuous: bio1 bio12 bio16 bio17 bio5 bio6 bio7 bio8 
#> Categorical: biome
#> [[4]]
#> Object of class SDMmodel 
#> Method: Maxnet 
#> 
#> Species: Virtual species 
#> Presence locations: 320 
#> Absence locations: 5000 
#> 
#> Model configurations:
#> --------------------
#> fc: lqph
#> reg: 2
#> 
#> Variables:
#> ---------
#> Continuous: bio1 bio12 bio16 bio17 bio5 bio6 bio7 bio8 
#> Categorical: biome
# Order rusults by highest test AUC
head(output@results[order(-output@results$test_AUC), ])
#>     fc reg train_AUC  test_AUC   diff_AUC
#> 2 lqph   1 0.8727191 0.8567250 0.01599406
#> 1  lqp   1 0.8634203 0.8519775 0.01144281
#> 4 lqph   2 0.8657109 0.8501575 0.01555344
#> 3  lqp   2 0.8597322 0.8486725 0.01105969

# Run the function using the AICc as metric and without saving the trained
# models, helpful when numerous hyperparameters are tested to avoid memory
# problems
output <- gridSearch(model, hypers = h, metric = "aicc", env = predictors,
                     save_models = FALSE)
#> Grid Search  ■■■■■■■■■■■■■                     40% | ETA:  5s - 00:00:3.1
#> Grid Search  ■■■■■■■■■■■■■■■■■■■               60% | ETA:  2s - 00:00:3.6
#> Grid Search  ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s - 00:00:5.3
#> 
output@results
#>     fc reg     AICc delta_AICc
#> 1  lqp   1 5276.603   7.797843
#> 2 lqph   1 5273.324   4.518590
#> 3  lqp   2 5283.850  15.044341
#> 4 lqph   2 5268.805   0.000000
# }