Maxent Variable Importance

Shows the percent contribution and permutation importance of the environmental variables used to train the model.

maxentVarImp(model)

Arguments

model
SDMmodel or SDMmodelCV object trained using the "Maxent" method.

Value

A data frame with the variable importance.

Details

When an SDMmodelCV object is passed to the function, the output is the average of the variable importance of each model trained during the cross validation.

See also

Examples

# \donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")

#> Extracting predictor information for presence locations...

#> Extracting predictor information for absence/background locations...

# Train a Maxent model
model <- train(method = "Maxent", data = data, fc = "l")

#> Error in .getMeVersion(): file missing:
#> /Users/runner/runners/2.165.2/work/_temp/Library/dismo/java/maxent.jar.
#> Please download it here: http://www.cs.princeton.edu/~schapire/maxent/

maxentVarImp(model)

#> Error in .get_model_class(model): object 'model' not found

# }