Model Report

Make a report that shows the main results.

modelReport(model, type, folder, test = NULL, response_curves = FALSE,
  jk = FALSE, env = NULL, clamp = TRUE, permut = 10)

Arguments

model
SDMmodel object.

type
character. Output type, see predict,Maxent-method for Maxent models or predict.maxnet for Maxnet models.

folder
character. The name of the folder in which to save the output. The folder is created in the working directory.

test
SWD object with the test locations, default is NULL.

response_curves
logical, if TRUE it plots the response curves in the html output, default is FALSE.

jk
logical, if TRUE it runs the jackknife test, default FALSE.

env
stack. If provided it computes and adds a prediction map to the output, default is NULL.

clamp
logical for clumping during prediction, used for response curves and for the prediction map, default is TRUE.

permut
integer. Number of permutations, default is 10.

Details

The function produces a report similar to the one created by MaxEnt software.

Examples

# NOT RUN {
modelReport(model, "cloglog", folder = "my_folder", response_curves = T)
# }