Model Report

Make a report that shows the main results.

modelReport(
  model,
  folder,
  test = NULL,
  type = NULL,
  response_curves = FALSE,
  only_presence = FALSE,
  jk = FALSE,
  env = NULL,
  clamp = TRUE,
  permut = 10
)

Arguments

model
SDMmodel object.

folder
character. The name of the folder in which to save the output. The folder is created in the working directory.

test
SWD object with the test locations, default is NULL.

type
character. The output type used for "Maxent" and "Maxnet" methods, possible values are "cloglog" and "logistic", default is NULL.

response_curves
logical, if TRUE it plots the response curves in the html output, default is FALSE.

only_presence
logical, if TRUE it uses only the range of the presence location for the marginal response, default is FALSE.

jk
logical, if TRUE it runs the jackknife test, default is FALSE.

env
stack. If provided it computes and adds a prediction map to the output, default is NULL.

clamp
logical for clumping during prediction, used for response curves and for the prediction map, default is TRUE.

permut
integer. Number of permutations, default is 10.

Details

The function produces a report similar to the one created by MaxEnt software.

Examples

# \donttest{
# If you run the following examples with the function example(), you may want
# to set the argument ask like following: example("modelReport", ask = FALSE)
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")

# Split presence locations in training (80%) and testing (20%) datasets
datasets <- trainValTest(data, test = 0.2, only_presence = TRUE)
train <- datasets[[1]]
test <- datasets[[2]]

# Train a model
model <- train(method = "Maxnet", data = train, fc = "lq")

# Create the report
if (FALSE) {
modelReport(model, type = "cloglog", folder = "my_folder", test = test,
            response_curves = TRUE, only_presence = TRUE, jk = TRUE,
            env = predictors, permut = 2)
}
# }