Optimize Model

The function uses a Genetic Algorithm implementation to optimize the model hyperparameter configuration according to the chosen metric.

Usage

optimizeModel(
  model,
  hypers,
  metric,
  test = NULL,
  pop = 20,
  gen = 5,
  env = NULL,
  keep_best = 0.4,
  keep_random = 0.2,
  mutation_chance = 0.4,
  interactive = TRUE,
  progress = TRUE,
  seed = NULL
)

Arguments

model

SDMmodel or SDMmodelCV object.

hypers

named list containing the values of the hyperparameters that should be tuned, see details.

metric

character. The metric used to evaluate the models, possible values are: "auc", "tss" and "aicc".

test

SWD object. Testing dataset used to evaluate the model, not used with aicc and SDMmodelCV objects.

pop

numeric. Size of the population.

gen

numeric. Number of generations.

env

stack containing the environmental variables, used only with "aicc".

keep_best

numeric. Percentage of the best models in the population to be retained during each iteration, expressed as decimal number.

keep_random

numeric. Probability of retaining the excluded models during each iteration, expressed as decimal number.

mutation_chance

numeric. Probability of mutation of the child models, expressed as decimal number.

interactive

logical, if FALSE the interactive chart is not created.

progress

logical, if TRUE shows a progress bar.

seed

numeric. The value used to set the seed to have consistent results.

Value

SDMtune object.

Details

To know which hyperparameters can be tuned you can use the output of the function getTunableArgs. Hyperparameters not included in the hypers argument take the value that they have in the passed model.

An interactive chart showing in real-time the steps performed by the algorithm is displayed in the Viewer pane.

Part of the code is inspired by this post.

See also

Author

Sergio Vignali

Examples

# \donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare presence and background locations
p_coords <- virtualSp$presence
bg_coords <- virtualSp$background

# Create SWD object
data <- prepareSWD(species = "Virtual species", p = p_coords, a = bg_coords,
                   env = predictors, categorical = "biome")
#> ℹ Extracting predictor information for presence locations
#> ✔ Extracting predictor information for presence locations [47ms]
#> 
#> ℹ Extracting predictor information for absence/background locations
#> ✔ Extracting predictor information for absence/background locations [100ms]
#> 

# Split presence locations in training (80%) and testing (20%) datasets
datasets <- trainValTest(data, val = 0.2, test = 0.2, only_presence = TRUE,
                         seed = 61516)
train <- datasets[[1]]
val <- datasets[[2]]

# Train a model
model <- train("Maxnet", data = train)

# Define the hyperparameters to test
h <- list(reg = seq(0.2, 5, 0.2),
          fc = c("l", "lq", "lh", "lp", "lqp", "lqph"))

# Run the function using as metric the AUC
if (FALSE) {
output <- optimizeModel(model, hypers = h, metric = "auc", test = val,
                        pop = 15, gen = 2, seed = 798)
output@results
output@models
output@models[[1]]  # Best model
}
# }