Thin Data

Remove all but one location per raster cell. The function removes NAs and if more than one location falls within the same raster cell it selects randomly one.

thinData(coords, env, x = "x", y = "y", progress = TRUE)

Arguments

coords

data.frame or matrix with the coordinates, see details.

env

stack containing the environmental variables, or a single raster layer.

x

character. Name of the column containing the x coordinates.

y

character. Name of the column containing the y coordinates.

progress

logical, if TRUE shows a progress bar.

Value

a matrix or a data frame with the thinned locations.

Details

coords and env must have the same coordinate reference system.

The coords argument can contain several columns. This is useful if the user has information related to the coordinates that doesn't want to loose with the thinning procedure. The function expects to have the x coordinates in a column named "x", and the y coordinates in a column named "y". If this is not the case, the name of the columns containing the coordinates can be specified using the arguments x and y.

Author

Sergio Vignali

Examples

# \donttest{
# Acquire environmental variables
files <- list.files(path = file.path(system.file(package = "dismo"), "ex"),
                    pattern = "grd", full.names = TRUE)
predictors <- raster::stack(files)

# Prepare background locations
bg_coords <- dismo::randomPoints(predictors, 9000)
nrow(bg_coords)
#> [1] 9000

# Thin the locations
# There are probably few coordinates that have NAs for some predictors, the
# function will remove these coordinates. Note that the function expects to
# the coordinates in two column named "x" and "y"
colnames(bg_coords)
#> [1] "x" "y"
thinned_bg <- thinData(bg_coords, env = predictors)
nrow(thinned_bg)
#> [1] 8993

# Here we double the coordinates and run the function again
thinned_bg <- thinData(rbind(bg_coords, bg_coords), env = predictors)
#> Thin Data  ■■■■■■■■■■■■■■■■■■■■              64% | ETA:  1s - 00:00:02
#> Thin Data  ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s - 00:00:3.2
#> 
nrow(thinned_bg)
#> [1] 8993

# In case of a dataframe containing more than two columns (e.g. a dataframe
# with the coordinates plus an additional column with the age of the species)
# and custom column names, use the function in this way
age <- sample(c(1, 2), size = nrow(bg_coords), replace = TRUE)
data <- cbind(age, bg_coords)
colnames(data) <- c("age", "X", "Y")
thinned_bg <- thinData(data, env = predictors, x = "X", y = "Y")
head(data)
#>      age      X      Y
#> [1,]   2 -68.25  -1.75
#> [2,]   1 -44.75 -11.75
#> [3,]   2 -78.25   0.25
#> [4,]   1 -75.25   8.25
#> [5,]   1 -46.75 -17.25
#> [6,]   1 -68.25 -34.25
# }